requirements.txt
Django==4.2
django-environ==0.9.0
djangorestframework~=3.14.0
python-dotenv~=1.0.0
assemblyai~=0.18.0
pydub~=0.25.1